Создал файл для хранения данных. Дописал функцию-шаблонизатор
