+++ data.php
@@ -1,0 +1,45 @@
+<?php
+// Массив категорий проектов
+$project_array=array("Все","Входящие","Учеба","Работа","Домашние дела","Авто");
+
+// Двумерных ассоциативный массив задач
+$array_tasks = [
+    'task1' => [
+        'task' => 'Собеседование в IT компании',
+        'date' => '01.06.2018',
+        'category' => 'Работа',
+        'complete' => 'Нет'
+    ],
+    'task2' => [
+        'task' => 'Выполнить тестовое задание',
+        'date' => '25.05.2018',
+        'category' => 'Работа',
+        'complete' => 'Нет'
+    ],
+    'task3' => [
+        'task' => 'Сделать задание первого раздела',
+        'date' => '21.04.2018',
+        'category' => 'Работа',
+        'complete' => 'Нет'
+    ],
+    'task4' => [
+        'task' => 'Встреча с другом',
+        'date' => '22.04.2018',
+        'category' => 'Учеба',
+        'complete' => 'Да'
+    ],
+    'task5' => [
+        'task' => 'Купить корм для кота',
+        'date' => 'нет',
+        'category' => 'Входящие',
+        'complete' => 'Нет'
+    ],
+    'task6' => [
+        'task' => 'Заказать пиццу',
+        'date' => 'нет',
+        'category' => 'Домашние дела',
+        'complete' => 'Нет'
+    ]
+];
+
+?>
